Moved extensions into a folder
