+++ Sources/SwiftGtk/Extensions/Bool+Conversions.swift
@@ -1,0 +1,9 @@
+import CGtk
+
+extension Bool {
+
+    func toGBoolean() -> gboolean {
+        self ? 1 : 0
+    }
+
+}
