+++ Sources/SwiftGtk/Extensions/GBoolean+Conversions.swift
@@ -1,0 +1,9 @@
+import CGtk
+
+extension gboolean {
+
+    func toBool() -> Bool {
+        self >= 1
+    }
+
+}
