+++ Sources/SwiftGtk/Extensions/Int+Conversions.swift
@@ -1,0 +1,9 @@
+import CGtk
+
+extension Int {
+
+    func toBool() -> Bool {
+        self >= 1
+    }
+
+}
