+++ Sources/SwiftGtk/Extensions/UnsafePointer+Conversions.swift
@@ -1,11 +1,9 @@
-//
-//  Copyright © 2017 Tomas Linhart. All rights reserved.
-//
-
 import CGtk
 
 extension UnsafePointer where Pointee == gchar {
+
     func toString() -> String {
-        return String(cString: self)
+        String(cString: self)
     }
+    
 }
