+++ Sources/SwiftGtk/Extensions/UnsafeRawPointer+Conversions.swift
@@ -1,0 +1,10 @@
+import CGtk
+
+extension UnsafeRawPointer {
+
+    func cast<T>() -> UnsafeMutablePointer<T> {
+        let pointer = UnsafeRawPointer(self).bindMemory(to: T.self, capacity: 1)
+        return UnsafeMutablePointer<T>(mutating: pointer)
+    }
+
+}
